+++ src/language/types.ts
@@ -35,7 +35,16 @@ export type SnippetParameters = Readonly<Record<ParameterTypeName, SnippetParame
 
 export type ParameterTypeMeta = Readonly<{ name: string; regexp: string }>
 
-export const LanguageNames = ['java', 'tsx', 'c_sharp', 'php', 'python', 'ruby', 'rust'] as const
+export const LanguageNames = [
+  'java',
+  'tsx',
+  'c_sharp',
+  'php',
+  'python',
+  'ruby',
+  'rust',
+  'javascript',
+] as const
 export type LanguageName = typeof LanguageNames[number]
 
 export type Source<L> = Readonly<{