+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Fixes a regression in the python language implementation for regexes
+
+### Fixed
+- Fixed a regression in the python language implementation for regexes [#119](https://github.com/cucumber/language-service/pull/119)
+
+### Added
+- Added support for JavaScript - [#42](https://github.com/cucumber/language-service/issues/42), [#115](https://github.com/cucumber/language-service/pull/115), [#120](https://github.com/cucumber/language-service/pull/120)
 
 ## [1.3.0] - 2022-11-28
 ### Added